attiny-hal: move eeprom definitions into mcu specific modules
It's quite hard to add new ATtiny devices to this crate. There are a
bunch of feature gates spread across the whole code base. In order to
get rid of them, move all mcu specific code into a single file. To keep
our current interface and avoid breaking existing user, the mcu specific
file defines a module for each periphal which gets reimported by our
periphal specific files.

No functional change intended.
