+++ mcu/attiny-hal/src/attiny2313.rs
@@ -1,0 +1,13 @@
+pub mod eeprom {
+    pub use crate::periphals::eeprom::*;
+
+    avr_hal_generic::impl_eeprom_attiny! {
+        hal: crate::Attiny,
+        peripheral: crate::pac::EEPROM,
+        capacity: 128,
+        addr_width: u8,
+        set_address: |peripheral, address| {
+            peripheral.eear.write(|w| w.bits(address));
+        },
+    }
+}
