+++ mcu/attiny-hal/src/attiny84.rs
@@ -3,3 +3,9 @@ pub mod adc {
 
     // Fixme: Implement ADC for ATtiny84.
 }
+
+pub mod eeprom {
+    pub use crate::periphals::eeprom::*;
+
+    // Fixme: Implement EEPROM for ATtiny84.
+}
