+++ mcu/attiny-hal/src/attiny88.rs
@@ -35,3 +35,17 @@ pub mod adc {
         },
     }
 }
+
+pub mod eeprom {
+    pub use crate::periphals::eeprom::*;
+
+    avr_hal_generic::impl_eeprom_attiny! {
+        hal: crate::Attiny,
+        peripheral: crate::pac::EEPROM,
+        capacity: 64,
+        addr_width: u8,
+        set_address: |peripheral, address| {
+            peripheral.eearl.write(|w| w.bits(address));
+        },
+    }
+}
